requirements.txt
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
psycopg2-binary==2.7.7
requests==2.21.0
urllib3==1.24.1
Werkzeug==0.14.1
